Finally have Astro setup again.
I accidentally updated Astronvim to a new version with very breaking
changes. The changes were so bad that I had to drop my old user config Dir and
start again.

This is the new starting place for my Astronvim user config.
I have all the plugins and custom key binds setup the way I like them
except for one(+1). j and k do not map to jzz and kzz as the new system
complains that jzz and kzz are unknown vairables. Even though they are
both string literals meant to be taking as a list of keystrokes.